removing classes from esi using direct calls to urls.

+++ model/Database/src/main/java/fr/guiguilechat/eveonline/model/database/yaml/Agent.java
@@ -2,7 +2,7 @@ package fr.guiguilechat.eveonline.model.database.yaml;
 
 public class Agent {
 
-	public int agentID;
+	public long agentID;
 	public String name;
 	public String agentType;
 	public String corporation;
